+++ src/Test/Unit/Model/Export/Catalog/ProductType/ConfigurableDataProviderTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Omikron\Factfinder\Model\Export\Catalog\ProductType;
 
-use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Catalog\Api\ProductRepositoryInterface;
 use Magento\Catalog\Model\Product;
 use Magento\ConfigurableProduct\Model\Product\Type\Configurable as ConfigurableProductType;
@@ -12,6 +11,7 @@ use Magento\Framework\Api\SearchCriteriaBuilder;
 use Omikron\Factfinder\Api\Filter\FilterInterface;
 use Omikron\Factfinder\Model\Export\Catalog\Entity\ProductVariationFactory;
 use Omikron\Factfinder\Model\Formatter\NumberFormatter;
+use Omikron\Factfinder\Test\TestHelper;
 use PHPUnit\Framework\TestCase;
 
 class ConfigurableDataProviderTest extends TestCase
@@ -24,7 +24,7 @@ class ConfigurableDataProviderTest extends TestCase
      */
     public function test_will_return_string_on_string_value()
     {
-        $valueOrEmptyStrMethod = $this->invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', ['test']);
+        $valueOrEmptyStrMethod = TestHelper::invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', ['test']);
         $this->assertEquals('test', $valueOrEmptyStrMethod);
     }
 
@@ -33,7 +33,7 @@ class ConfigurableDataProviderTest extends TestCase
      */
     public function test_will_return_empty_string_on_null_value()
     {
-        $valueOrEmptyStrMethod = $this->invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', [null]);
+        $valueOrEmptyStrMethod = TestHelper::invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', [null]);
         $this->assertEquals('', $valueOrEmptyStrMethod);
     }
 
@@ -42,7 +42,7 @@ class ConfigurableDataProviderTest extends TestCase
      */
     public function test_will_return_empty_string_on_bool_value()
     {
-        $valueOrEmptyStrMethod = $this->invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', [false]);
+        $valueOrEmptyStrMethod = TestHelper::invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', [false]);
         $this->assertEquals('', $valueOrEmptyStrMethod);
     }
 
@@ -51,16 +51,20 @@ class ConfigurableDataProviderTest extends TestCase
      */
     public function test_will_return_empty_string_on_array_value()
     {
-        $valueOrEmptyStrMethod = $this->invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', [[]]);
+        $valueOrEmptyStrMethod = TestHelper::invokeMethod($this->configurableDataProvider, 'valueOrEmptyStr', [[]]);
         $this->assertEquals('', $valueOrEmptyStrMethod);
     }
 
-    public function invokeMethod(&$object, $methodName, array $parameters = [])
+    /**
+     * @covers ConfigurableDataProvider::getChildren
+     */
+    public function test_will_no_throw_error_if_there_is_no_chlidren_ids()
     {
-        $reflection = new \ReflectionClass(get_class($object));
-        $method = $reflection->getMethod($methodName);
-        $method->setAccessible(true);
-        return $method->invokeArgs($object, $parameters);
+        $this->productMock->method('getId')->willReturn('1');
+        $this->configurableProductTypeMock->method('getChildrenIds')->with('1')
+            ->willReturn([]);
+        $variants = TestHelper::invokeMethod($this->configurableDataProvider, 'getChildren', [$this->productMock]);
+        $this->assertEquals([], $variants);
     }
 
     protected function setUp(): void